:truck: Move Consts
